Move MapServer to pkg.
